refactor(JSON-RPC): move code around
