+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -6,6 +6,7 @@ mod broadcasted_transaction_test;
 pub mod deprecated_contract_class;
 #[cfg(test)]
 mod deprecated_contract_class_test;
+pub mod error;
 pub mod state;
 pub mod transaction;
 #[cfg(test)]
